fix gcc compile problems

+++ js_dbindex.h
@@ -39,7 +39,7 @@ typedef struct {
 	//	field name bytes follow
 
 	struct Field {
-		uint64_t hash;		// field name hash value
+		uint32_t hash;		// field name hash value
 		uint32_t len[1];	// length of field name
 		uint8_t name[1];	// field name
 	} field;
